+++ .github/scripts/comment-generator.js
@@ -7,6 +7,7 @@
 import fs from 'fs';
 import path from 'path';
 import process from 'process';
+import { fileURLToPath } from 'url';
 
 /**
  * Normalize broken link data from different report formats
@@ -316,7 +317,7 @@ Examples:
 }
 
 // Run CLI if this file is executed directly
-if (import.meta.url === `file://${process.argv[1]}`) {
+if (fileURLToPath(import.meta.url) === process.argv[1]) {
   main();
 }
